separate booking View file and tidy up controller
