+++ static/js/controller/booking.js
@@ -1,4 +1,5 @@
 import * as View from "../view/view.js";
+import * as BookingView from "../view/booking.js";
 
 const trashBtn = document.querySelector(".trash");
 
@@ -47,52 +48,27 @@ export async function fetchGetBooking() {
     });
 
     if (!response.ok) {
-      removeDisplayBooking();
-      noBooking();
-      displayName(userName);
+      BookingView.removeDisplayBooking();
+      View.setElementDisplay(".result", "flex");
+      BookingView.displayUserName(userName);
       console.log("Failed to fetch booking details:", response.status);
       return;
     }
 
     const data = await response.json();
     if (!data || !data.data) {
-      removeDisplayBooking();
-      noBooking();
-      displayName(userName);
+      BookingView.removeDisplayBooking();
+      View.setElementDisplay(".result", "flex");
+      BookingView.displayUserName(userName);
       console.error("No booking data available");
       return;
     }
     console.log(data.data);
-    displayBooking(data.data, userName);
-    displayName(userName);
+    BookingView.updateBookingDetails(data.data, userName);
+    BookingView.displayUserName(userName);
   } catch (error) {
     console.error("Error fetching attraction:", error);
   }
-}
-function displayName(userName) {
-  const title = document.querySelector(".no-booking-title");
-  title.textContent = `您好，${userName}，待預定的行程如下：`;
-}
-
-function displayBooking(attraction) {
-  const image = document.querySelector(".fade");
-  const attractionName = document.querySelector(".attraction-name");
-  const date = document.querySelector(".date-detail");
-  const time = document.querySelector(".time-detail");
-  const price = document.querySelector(".price-detail");
-  const address = document.querySelector(".address-detail");
-  const totalPrice = document.querySelector(".booking-price");
-
-  image.src = attraction.attraction.image;
-  attractionName.textContent = `台北一日遊：${attraction.attraction.name}`;
-  date.textContent = attraction.date;
-  time.textContent = attraction.time;
-  price.textContent = attraction.price;
-  address.textContent = attraction.attraction.address;
-  totalPrice.textContent = `總價：新台幣${attraction.price}`;
-
-  time.textContent =
-    attraction.time === "上半天" ? "上午9點到下午4點" : "下午5點到晚上12點";
 }
 
 async function fetchDeleteBooking() {
@@ -110,29 +86,9 @@ async function fetchDeleteBooking() {
       return;
     }
 
-    removeDisplayBooking();
+    BookingView.removeDisplayBooking();
     window.location.reload();
   } catch (error) {
     console.error("Error deleting attraction:", error);
   }
-}
-
-function removeDisplayBooking() {
-  const bookingDetail = document.querySelector(".booking-location");
-  const seperates = document.querySelectorAll(".seperate");
-  const contactArea = document.querySelector(".contact-area");
-  const cardArea = document.querySelector(".card-area");
-  const totalSubmit = document.querySelector(".total-submit");
-
-  bookingDetail.remove();
-  contactArea.remove();
-  cardArea.remove();
-  totalSubmit.remove();
-  seperates.forEach((seperate) => {
-    if (seperate) seperate.remove();
-  });
-}
-
-function noBooking() {
-  View.setElementDisplay(".result", "flex");
 }
